requirements.txt
requests==2.31.0
sentence_transformers==2.7.0
pinecone-client==3.2.2
langchain==0.1.16
flask==3.0.3
beautifulsoup4==4.12.3
bs4==0.0.2 
huggingface-hub==0.22.2 
numpy==1.26.4 
python-dotenv==1.0.1 
scikit-learn==1.4.2 
sentence-transformers==2.7.0 
transformers==4.39.3
openai==1.21.0